Implimenting gate

+++ app/Http/Controllers/RoleController.php
@@ -9,12 +9,14 @@ use App\Http\Resources\RoleControllerResource;
 use App\Models\Permission;
 use App\Models\Role;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Inertia\Inertia;
 
 class RoleController extends Controller
 {
    public function index()
     {
+        Gate::authorize('role_access');
         $roles = Role::with('permissions')->get();
         return Inertia::render('Roles/Index', [
             'roles' => RoleControllerResource::collection($roles),
@@ -23,6 +25,8 @@ class RoleController extends Controller
 
     public function create()
     {
+        Gate::authorize('role_create');
+
         $permissions = Permission::all();
         return Inertia::render('Roles/Create', [
             'permissions' => PersmissionControllerResource::collection($permissions),
@@ -30,6 +34,8 @@ class RoleController extends Controller
     }
 
     public function store(StoreRoleRequest $request){
+        Gate::authorize('role_store');
+
         $permissions = collect($request->selectedPermission)->map(function($permissions){
             return $permissions['value'];
         });
@@ -44,6 +50,7 @@ class RoleController extends Controller
     }
 
      public function edit(Role $role){
+        Gate::authorize('role_edit');
         $permissions = Permission::all();
         return Inertia::render('Roles/Edit',[
             'role' => RoleControllerResource::make($role->load('permissions')),
@@ -52,6 +59,7 @@ class RoleController extends Controller
     }
 
     public function update(UpdateRoleRequest $request, Role $role){
+        Gate::authorize('role_update');
         $role->update($request->validated());
         $permissions = collect($request->selectedPermission)->map(function($permissions){
             return $permissions['value'];
@@ -66,6 +74,7 @@ class RoleController extends Controller
 
     public function destroy(Role $role)
     {
+        Gate::authorize('role_delete');
         $role->delete();
 
         return redirect()->route('roles.index')->with('message', [
